refactor; minor fixes; remove dup tests
